// src/app/models/administration/company.model.ts
import { Taxe } from './taxe.model';
import { Resource } from '../shared/ressource.model';
import { FileInfo } from '../shared/objectToSend';
import { PurchaseSettings } from '../purchase/purchase-settings.model';
import { SaleSettings } from '../sales/sales-settings.model';
import { Contact } from '../shared/contact.model';
import { Currency } from './currency.model';
import { Address } from '../crm/address.model';

export class Company extends Resource {
  Code: string;
  MatriculeFisc: string;
  Name: string;
  SharedDocumentsKey: string;
  Description: string;
  CommercialRegister: string;
  TaxNumber: string;
  Email: string;
  WebSite: string;
  Logo: string;
  Adress: string;
  Tel1: string;
  Tel2: string;
  Fax: string;
  IdCurrency: number;
  Culture: string;
  TaxIdentNumber: string;
  Siret: string;
  Region: string;
  IdCountry: number;
  IdCity: number;
  DaysOfWork: number;
  PayDependOnTimesheet: boolean;
  FiscalStamp: number;
  Nic: string;
  IdNaf: number;
  Capital: number;
  PaymentOffset: string;
  IdAtrate: number;
  HeuRef: number;
  RegularisationMode: string;
  AttachmentUrl: string;
  PictureFileInfo: FileInfo;
  VatNumber: string;
  CnssAffiliation: string;
  PurchaseSettings: PurchaseSettings;
  SaleSettings: SaleSettings;
  Contact: Array<Contact>;
  IdAccountingAccount: number;
  TimeSheetPerHalfDay: boolean;
  IdCurrencyNavigation: Currency;
  Category: string;
  SecondaryEstablishment: string;
  WithholingTax: boolean;
  Address: Address[];
  ActivityArea: number;
  Timezone: string;
  AllowEditionItemDesignation: boolean;
  DaysOfWeekWorked: Array<number>;
  IdDefaultTax: number;
  IdDefaultTaxNavigation: Taxe;
  AllowRelationSupplierItems: boolean;
  NoteIsRequired: boolean;

}